% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubicar.R
\name{cubicar}
\alias{cubicar}
\title{Calcula el cubo de un valor}
\usage{
cubicar(x)
}
\arguments{
\item{x}{un \code{numeric()} o \code{integer()} con el valor inicial. Puede ser un
vector, en cuyo caso se usaran las reglas de reciclaje de R.}
}
\value{
El cubo del valor \code{x}
}
\description{
Calcula el cubo de un valor
}
\examples{
cubicar(10)
}
